third_party/WebKit/LayoutTests/external/wpt/webrtc/RTCSctpTransport-constructor-expected.txt

This is a testharness.js-based test.
FAIL setRemoteDescription() with answer containing data media should initialize pc.sctp assert_equals: expected (object) null but got (undefined) undefined
FAIL setLocalDescription() with answer containing data media should initialize pc.sctp assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
